move RepoOptions to internal/search
