+++ internal/search/repos/repos_test.go
@@ -164,7 +164,7 @@ func TestRevisionValidation(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.repoFilters[0], func(t *testing.T) {
-			op := Options{RepoFilters: tt.repoFilters}
+			op := search.Options{RepoFilters: tt.repoFilters}
 			repositoryResolver := &Resolver{}
 			resolved, err := repositoryResolver.Resolve(context.Background(), op)
 
@@ -431,7 +431,7 @@ func TestUseIndexableReposIfMissingOrGlobalSearchContext(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			op := Options{
+			op := search.Options{
 				SearchContextSpec: tt.searchContextSpec,
 				Query:             queryInfo,
 			}
@@ -507,7 +507,7 @@ func TestResolveRepositoriesWithUserSearchContext(t *testing.T) {
 		database.Mocks.Namespaces.GetByName = nil
 	}()
 
-	op := Options{
+	op := search.Options{
 		Query:             queryInfo,
 		SearchContextSpec: "@" + wantName,
 	}
@@ -588,7 +588,7 @@ func TestResolveRepositoriesWithSearchContext(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	op := Options{
+	op := search.Options{
 		Query:             queryInfo,
 		SearchContextSpec: "searchcontext",
 	}
